test: PlayerName 명령어와 같은 이름 예외 발생 테스트 추가
